style: add transition effect, upd naming
