+++ packages/client/src/ui/avatarEl/style.module.scss
@@ -1,0 +1,39 @@
+.avatar {
+  position: relative;
+  width: 100px;
+  height: 100px;
+  border-radius: 50%;
+  cursor: pointer;
+
+  &:hover .wrapper {
+    transform: rotateY(180deg);
+  }
+}
+
+.wrapper {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  transition: transform 300ms;
+  transform-style: preserve-3d;
+}
+
+.img,
+.back {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0px;
+  left: 0px;
+  border-radius: 50%;
+  backface-visibility: hidden;
+}
+
+.back {
+  display: flex;
+  align-items: center;
+  text-align: center;
+  color: var(--text);
+  background: var(--shadow-light);
+  transform: rotateY(180deg);
+}
