--- packages/client/src/ui/avatarImg/style.module.scss
@@ -1,20 +1,0 @@
-.wrapper {
-  position: relative;
-  display: flex;
-  width: 100px;
-  height: 100px;
-  border: 2px solid blue;
-  border-radius: 50%;
-  background-color: gray;
-  cursor: pointer;
-}
-
-.img {
-  position: relative;
-  width: 100%;
-  height: 100%;
-  top: 0px;
-  left: 0px;
-  border-radius: 50%;
-  background-color: green;
-}
